feat: introduce external knowledge frontend (#82)
* initial styling shared db's feature

* component externaldb

* WIP: introducing Redux states

* introduce redux store for external db connection

* first draft frontend shared db

* adjust Config page with apiCall middleware


Left the changes from old version in comments just to checkback. Can be deleted once it's fully rolled out. 
Introduced a state `connectedExternalDb` which will be checked in the middleware to define the api call.

Once the backend is done it needs some adjustment's in the calls to send api/keys,..